Ygdra.Python/requirements.txt
pip
wheel
pyspark
pytest
pytest-runner
pytest-cov
azure-identity
pandas
IPython
ipykernel
pandas
